frontend/Backend/requirement.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
Markdown==3.7
sqlparse==0.5.3
typing_extensions==4.12.2
asgiref==3.8.1
Django==5.1.5
django-jazzmin==3.0.1
djangorestframework==3.15.2
Markdown==3.7
sqlparse==0.5.3
typing_extensions==4.12.2
